Fix atime/mtime/ctime output of stat(1)
stat(1) calls human_time() for converting a unix timestamp to a human
readable format. human_time() however returns a pointer to a static
buffer so it must not be called more than once in an argument list (as
it used to be for printf()); it will evaluate to the same value anyway.

Change-Id: I6e4dcf33f1198e6c974f66214d6abe68818e4465